refactor: organise src/components into domain subfolders
src/components/ gains audio/, debug/, engine/, input/, menu/, pause/,
portal/, scene/, streaming/ folders alongside the existing settings/
and ui/. MenuSettingsView moves into settings/ (it's the tabs
container) and Sparkline moves into ui/. GooseMode, I18nSync, and
WindowControls stay at the root as cross-cutting infra.

All file moves use git mv so history is preserved.
