+++ src/components/debug/InputOverlay.tsx
@@ -1,10 +1,10 @@
 import { useEffect, useMemo, useRef, useState } from 'react'
-import { useConnection } from '../context/streaming/connection'
-import { useInput } from '../context/streaming/input'
-import { useSettings } from '../hooks/settings/settingsContextValue'
-import type { InputCode } from '../types/input'
-import { CODE_MAP } from '../hooks/input/useGameInput'
-import VirtualGamepad, { type GamepadAxes } from './VirtualGamepad'
+import { useConnection } from '../../context/streaming/connection'
+import { useInput } from '../../context/streaming/input'
+import { useSettings } from '../../hooks/settings/settingsContextValue'
+import type { InputCode } from '../../types/input'
+import { CODE_MAP } from '../../hooks/input/useGameInput'
+import VirtualGamepad, { type GamepadAxes } from '../input/VirtualGamepad'
 
 // QWERTY keyboard layout (simple labels) — copied verbatim from owl-tube/app/InputDisplay/constants.ts
 const KEYBOARD_LAYOUT = [
