+++ src/components/engine/TerminalDisplay.tsx
@@ -1,20 +1,20 @@
 import { useCallback, useEffect, useMemo, useState } from 'react'
-import { invoke } from '../bridge'
-import type { TranslationKey } from '../i18n'
-import { buildDiagnosticsPayload } from '../lib/diagnosticsPayload'
-import { resolveStage } from '../stages'
-import { useConnection } from '../context/streaming/connection'
-import { useWebsocket } from '../context/streaming/websocket'
-import { useVortex } from '../context/vortex/vortexContextValue'
-import { useSettings } from '../hooks/settings/settingsContextValue'
-import { useEngineLogs } from '../hooks/engine/useEngineLogs'
-import Button from './ui/Button'
-import { GooseFactTicker } from './GooseMode'
-import { isGooseMode } from '../i18n'
-import RawButton from './ui/RawButton'
+import { invoke } from '../../bridge'
+import type { TranslationKey } from '../../i18n'
+import { buildDiagnosticsPayload } from '../../lib/diagnosticsPayload'
+import { resolveStage } from '../../stages'
+import { useConnection } from '../../context/streaming/connection'
+import { useWebsocket } from '../../context/streaming/websocket'
+import { useVortex } from '../../context/vortex/vortexContextValue'
+import { useSettings } from '../../hooks/settings/settingsContextValue'
+import { useEngineLogs } from '../../hooks/engine/useEngineLogs'
+import Button from '../ui/Button'
+import { GooseFactTicker } from '../GooseMode'
+import { isGooseMode } from '../../i18n'
+import RawButton from '../ui/RawButton'
 import ServerLogDisplay from './ServerLogDisplay'
-import SocialCtaRow from './SocialCtaRow'
-import { FocusScope } from '../context/focus/FocusScopeContext'
+import SocialCtaRow from '../menu/SocialCtaRow'
+import { FocusScope } from '../../context/focus/FocusScopeContext'
 import { useTranslation } from 'react-i18next'
 
 const INLINE_ERROR_MAX_LENGTH = 80
