+++ src/components/input/VirtualGamepad.tsx
@@ -3,8 +3,8 @@
  *  the live stick deflections (polled directly from `navigator.getGamepads()`).
  *  Fades out after a few seconds of inactivity. */
 
-import { GAMEPAD_CODES } from '../hooks/input/useGameInput'
-import type { InputCode } from '../types/input'
+import { GAMEPAD_CODES } from '../../hooks/input/useGameInput'
+import type { InputCode } from '../../types/input'
 
 const U = 4 // matches VirtualKeyboard/VirtualMouse unit (cqh)
