+++ src/components/menu/PortalPreview.tsx
@@ -1,7 +1,7 @@
 import { useRef, type CSSProperties, type ReactNode } from 'react'
-import { usePortalMediaMount } from '../hooks/portal/usePortalMediaMount'
-import PortalSparks from './PortalSparks'
-import { PORTAL_ENTER_DURATION_MS, PORTAL_SHRINK_DURATION_MS, PORTAL_SHRINK_END_SCALE } from '../lib/portalAnimation'
+import { usePortalMediaMount } from '../../hooks/portal/usePortalMediaMount'
+import PortalSparks from '../portal/PortalSparks'
+import { PORTAL_ENTER_DURATION_MS, PORTAL_SHRINK_DURATION_MS, PORTAL_SHRINK_END_SCALE } from '../../lib/portalAnimation'
 
 type PortalPreviewProps = {
   videoElement: HTMLVideoElement | null
