+++ src/components/pause/PauseMainView.tsx
@@ -1,15 +1,15 @@
 import { useRef, useState, type ChangeEvent, type DragEvent } from 'react'
-import type { SeedRecord } from '../types/app'
-import SceneGrid from './SceneGrid'
-import SceneCardAddButton from './SceneCardAddButton'
-import SceneAuthoringPrompt from './SceneAuthoringPrompt'
-import SocialCtaRow from './SocialCtaRow'
-import MenuButton from './ui/MenuButton'
-import Slider from './ui/Slider'
-import { VIEW_DESCRIPTION, VIEW_HEADING } from '../styles'
-import { ALLOW_USER_SCENES } from '../constants'
+import type { SeedRecord } from '../../types/app'
+import SceneGrid from '../scene/SceneGrid'
+import SceneCardAddButton from '../scene/SceneCardAddButton'
+import SceneAuthoringPrompt from '../scene/SceneAuthoringPrompt'
+import SocialCtaRow from '../menu/SocialCtaRow'
+import MenuButton from '../ui/MenuButton'
+import Slider from '../ui/Slider'
+import { VIEW_DESCRIPTION, VIEW_HEADING } from '../../styles'
+import { ALLOW_USER_SCENES } from '../../constants'
 import { useTranslation } from 'react-i18next'
-import { useSettings } from '../hooks/settings/settingsContextValue'
+import { useSettings } from '../../hooks/settings/settingsContextValue'
 
 interface PauseMainViewProps {
   scenes: SeedRecord[]
