+++ src/components/pause/PauseOverlay.tsx
@@ -1,21 +1,21 @@
 import { useEffect, useMemo, useState } from 'react'
 import { AnimatePresence, motion } from 'framer-motion'
-import { useInput } from '../context/streaming/input'
-import { useSeeds } from '../context/streaming/seeds'
-import { useSession } from '../context/streaming/session'
-import { useWebsocket } from '../context/streaming/websocket'
-import MenuSettingsView from './MenuSettingsView'
+import { useInput } from '../../context/streaming/input'
+import { useSeeds } from '../../context/streaming/seeds'
+import { useSession } from '../../context/streaming/session'
+import { useWebsocket } from '../../context/streaming/websocket'
+import MenuSettingsView from '../settings/MenuSettingsView'
 import PauseMainView from './PauseMainView'
-import { PAUSE_VIEW, type PauseViewKey } from '../constants'
-import { viewFadeVariants } from '../transitions'
-import { useSeedManager } from '../hooks/seeds/useSeedManager'
-import { useSceneOrder } from '../hooks/scene/useSceneOrder'
-import { usePointerLockFeedback } from '../hooks/input/usePointerLockFeedback'
-import { useSceneActions } from '../hooks/scene/useSceneActions'
-import { useSceneGeneration } from '../hooks/scene/useSceneGeneration'
-import type { SeedRecord } from '../types/app'
-import { useSettings } from '../hooks/settings/settingsContextValue'
-import { FocusScope } from '../context/focus/FocusScopeContext'
+import { PAUSE_VIEW, type PauseViewKey } from '../../constants'
+import { viewFadeVariants } from '../../transitions'
+import { useSeedManager } from '../../hooks/seeds/useSeedManager'
+import { useSceneOrder } from '../../hooks/scene/useSceneOrder'
+import { usePointerLockFeedback } from '../../hooks/input/usePointerLockFeedback'
+import { useSceneActions } from '../../hooks/scene/useSceneActions'
+import { useSceneGeneration } from '../../hooks/scene/useSceneGeneration'
+import type { SeedRecord } from '../../types/app'
+import { useSettings } from '../../hooks/settings/settingsContextValue'
+import { FocusScope } from '../../context/focus/FocusScopeContext'
 
 const PauseOverlayContent = () => {
   const requestPointerLock = useInput().pointerLock.request
