+++ src/components/scene/SceneCardAddButton.tsx
@@ -1,4 +1,4 @@
-import { useUISound } from '../hooks/audio/useUISound'
+import { useUISound } from '../../hooks/audio/useUISound'
 import SceneCardBase from './SceneCardBase'
 
 interface SceneCardAddButtonProps {
