+++ src/components/scene/SceneCardBase.tsx
@@ -1,5 +1,5 @@
 import type { DragEvent, ReactNode } from 'react'
-import { useUISound } from '../hooks/audio/useUISound'
+import { useUISound } from '../../hooks/audio/useUISound'
 
 type DragButtonHandler = (event: DragEvent<HTMLButtonElement>) => void
