+++ src/components/settings/MenuSettingsView.tsx
@@ -1,24 +1,24 @@
 import { useState, useEffect, useCallback, useRef } from 'react'
 import { useTranslation } from 'react-i18next'
-import type { TranslationKey } from '../i18n'
-import { VIEW_DESCRIPTION, VIEW_HEADING } from '../styles'
-import { useSettings } from '../hooks/settings/settingsContextValue'
-import { useConnection } from '../context/streaming/connection'
-import { ENGINE_MODES } from '../types/settings'
-import { useVolumeControls } from '../hooks/audio/useVolumeControls'
-import MenuButton from './ui/MenuButton'
-import SettingsToggle from './ui/SettingsToggle'
-import { useGamepadConnected } from '../hooks/input/useGameInput'
-import { FocusScope } from '../context/focus/FocusScopeContext'
-import Modal from './ui/Modal'
-import ConfirmModal from './ui/ConfirmModal'
-import Button from './ui/Button'
-import attributionText from '../../assets/audio/ATTRIBUTION.md?raw'
-import GeneralTab from './settings/GeneralTab'
-import EngineTab, { type EngineTabHandle } from './settings/EngineTab'
-import KeyboardTab, { type KeyboardTabHandle } from './settings/KeyboardTab'
-import GamepadTab, { type GamepadTabHandle } from './settings/GamepadTab'
-import DebugTab, { type DebugTabHandle } from './settings/DebugTab'
+import type { TranslationKey } from '../../i18n'
+import { VIEW_DESCRIPTION, VIEW_HEADING } from '../../styles'
+import { useSettings } from '../../hooks/settings/settingsContextValue'
+import { useConnection } from '../../context/streaming/connection'
+import { ENGINE_MODES } from '../../types/settings'
+import { useVolumeControls } from '../../hooks/audio/useVolumeControls'
+import MenuButton from '../ui/MenuButton'
+import SettingsToggle from '../ui/SettingsToggle'
+import { useGamepadConnected } from '../../hooks/input/useGameInput'
+import { FocusScope } from '../../context/focus/FocusScopeContext'
+import Modal from '../ui/Modal'
+import ConfirmModal from '../ui/ConfirmModal'
+import Button from '../ui/Button'
+import attributionText from '../../../assets/audio/ATTRIBUTION.md?raw'
+import GeneralTab from './GeneralTab'
+import EngineTab, { type EngineTabHandle } from './EngineTab'
+import KeyboardTab, { type KeyboardTabHandle } from './KeyboardTab'
+import GamepadTab, { type GamepadTabHandle } from './GamepadTab'
+import DebugTab, { type DebugTabHandle } from './DebugTab'
 
 type MenuSettingsViewProps = {
   onBack: () => void
