+++ src/components/streaming/ConnectionLostOverlay.tsx
@@ -1,10 +1,10 @@
 import { useCallback } from 'react'
-import { buildDiagnosticsPayload } from '../lib/diagnosticsPayload'
-import { useConnection } from '../context/streaming/connection'
-import { useWebsocket } from '../context/streaming/websocket'
-import { useSettings } from '../hooks/settings/settingsContextValue'
-import Button from './ui/Button'
-import ServerLogDisplay from './ServerLogDisplay'
+import { buildDiagnosticsPayload } from '../../lib/diagnosticsPayload'
+import { useConnection } from '../../context/streaming/connection'
+import { useWebsocket } from '../../context/streaming/websocket'
+import { useSettings } from '../../hooks/settings/settingsContextValue'
+import Button from '../ui/Button'
+import ServerLogDisplay from '../engine/ServerLogDisplay'
 import { useTranslation } from 'react-i18next'
 
 const MODAL_BUTTON = 'p-[0.5cqh_1.78cqh] text-[2.49cqh]'
